Ensure strict namespacing of linker symbols.
Whenever a functions defined in some standard calls a function not
defined in that standard, we need to make sure that the call is done
through a symbol reserved by the standard. Also we need to make sure
that the called function also does not make use of unreserved symbols.
This is because programs conforming to a standard may define any symbols
that are not reserved by that standard, and expect that the
implementation still works.

Files in libc now contain a comment that states which standard defines
the given interface and whether it is called from interfaces from other
standards. Only files that had to be changed include these new comments.